pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "orchestra-knowledge-capture"
version = "0.1.0"
description = "Pipe external knowledge into your Orchestra wiki — PDFs, URLs, text, and research topics become a queryable knowledge graph."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["knowledge-base", "wiki", "rag", "ai", "orchestra"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "jinja2>=3.1.4",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    "openai>=1.30.0",
    "pyyaml>=6.0.1",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "pypdf>=4.2.0",
    "trafilatura>=1.9.0",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.0",
]

[project.scripts]
knowledge-capture = "server:main"

[project.urls]
Homepage = "https://github.com/randomchaos7800-hub/orchestra-knowledge-capture"
Repository = "https://github.com/randomchaos7800-hub/orchestra-knowledge-capture"
"Orchestra" = "https://github.com/randomchaos7800-hub/orchestra"
